v5.7.2

+++ src/main/java/bot/antony/commands/aam/events/ProposalListener.java
@@ -1,5 +1,7 @@
 package bot.antony.commands.aam.events;
 
+import java.util.Arrays;
+
 import bot.antony.Antony;
 import net.dv8tion.jda.api.entities.ChannelType;
 import net.dv8tion.jda.api.entities.Guild;
@@ -12,30 +14,24 @@ public class ProposalListener extends ListenerAdapter {
 
 	@Override
 	public void onMessageReceived(MessageReceivedEvent event) {
-				
-		//check which channel ...
 		if(event.isFromType(ChannelType.TEXT)) {
-			if(event.getMessage().getType() != MessageType.CHANNEL_PINNED_ADD) {
-				final TextChannel channel = event.getTextChannel();
-				Guild guild = event.getGuild();
+			Guild guild = event.getGuild();
+			TextChannel proposalChan = Antony.getGuildController().getValidChannel(guild, Arrays.asList(650687328863518740L, 778960515895918632L)); //Prod, Test
+			
+			//check which channel ...
+			if(event.getTextChannel() == proposalChan) {
 				
-				//TODO: Use Variable
-				long proposalChannelId = 778960515895918632L;	//is test ID
-				if(Antony.isProdStage()) {
-					proposalChannelId = 650687328863518740L;
-				}
-				
-				//Is a new offer
-				if(channel.getIdLong() == proposalChannelId) {				
-					//guild.getTextChannelById(proposalChannelId).sendMessage(userString).queue();
+				if(event.getMessage().getType() != MessageType.CHANNEL_PINNED_ADD) {
+	
 					if(guild.getEmotesByName("ausstehend", true).size() > 0) {
 						event.getMessage().addReaction(guild.getEmotesByName("ausstehend", true).get(0)).complete();
 					}
 					event.getMessage().pin().complete();
+					
+				} else {
+					event.getMessage().delete().complete();
 				}
-			} else {
-				event.getMessage().delete().complete();
 			}
-		} 
+		}
 	}
 }
